Move non-implementation-specific action handling logic from DefaultActivityActionHandlers.kt to new object DefaultAndroidActionHandlers.kt
